@@ -1,10 +1,11 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
 
-Vue.config.productionTip = false;
+import "./assets/basic.less";
+import "./assets/content.css";
+import "./assets/reset.min.css";
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount("#app");
+createApp(App)
+  .use(router)
+  .mount("#app");
