@@ -1,18 +1,32 @@
 <template>
   <div class="home">
-    <img alt="Vue logo" src="../assets/logo.png" />
-    <HelloWorld msg="Welcome to Your Vue.js App" />
+    <zhihu-header />
+    <zhihu-banner />
   </div>
 </template>
 
 <script>
 // @ is an alias to /src
-import HelloWorld from "@/components/HelloWorld.vue";
+import zhihuHeader from "@/components/Header.vue";
+import zhihuBanner from "@/components/Banner.vue";
+import { onBeforeMount,reactive } from 'vue';
+import { zhihu } from '../api/index';
 
 export default {
   name: "Home",
   components: {
-    HelloWorld
+    zhihuHeader,
+    zhihuBanner
+  },
+  setup(){
+    const state = reactive({
+      date: new Date(),
+      bannerList: [],
+
+    })
+     onBeforeMount(() => {
+      const data = zhihu.API_LATEST
+    });
   }
 };
 </script>
